wip try to collapse Resolve into one impl
